Crate notmuch_sys

Not much of an email library, just index and search.

Structs

notmuch_database_t

notmuch_directory_t

notmuch_filenames_t

notmuch_message_t

notmuch_messages_t

notmuch_query_t

notmuch_tags_t

notmuch_thread_t

notmuch_threads_t

Enums

notmuch_database_mode_t
Database open mode for notmuch_database_open.

notmuch_exclude_t
Exclude values for notmuch_query_set_omit_excluded

notmuch_message_flag_t
Message flags.

notmuch_sort_t
Sort values for notmuch_query_set_sort.

notmuch_status_t
Status codes used for the return values of most functions.

Constants

TAG_MAX
The longest possible tag value.

Functions

notmuch_database_add_message
Add a new message to the given notmuch database or associate an additional filename with an existing message.

notmuch_database_begin_atomic
Begin an atomic database operation.

notmuch_database_close
Commit changes and close the given notmuch database.

notmuch_database_compact
Compact a notmuch database, backing up the original database to the given path.

notmuch_database_create
Create a new, empty notmuch database located at 'path'.

notmuch_database_create_verbose
Like notmuch_database_create, except optionally return an error message. This message is allocated by malloc and should be freed by the caller.

notmuch_database_destroy
Destroy the notmuch database, closing it if necessary and freeing all associated resources.

notmuch_database_end_atomic
Indicate the end of an atomic database operation.

notmuch_database_find_message
Find a message with the given message_id.

notmuch_database_find_message_by_filename
Find a message with the given filename.

notmuch_database_get_all_tags
Return a list of all tags found in the database.

notmuch_database_get_directory
Retrieve a directory object from the database for 'path'.

notmuch_database_get_path
Return the database path of the given database.

notmuch_database_get_revision
Return the committed database revision and UUID.

notmuch_database_get_version
Return the database format version of the given database.

notmuch_database_needs_upgrade
Can the database be upgraded to a newer database version?

notmuch_database_open
Open an existing notmuch database located at 'path'.

notmuch_database_open_verbose
Like notmuch_database_open, except optionally return an error message. This message is allocated by malloc and should be freed by the caller.

notmuch_database_remove_message
Remove a message filename from the given notmuch database. If the message has no more filenames, remove the message.

notmuch_database_status_string
Retrieve last status string for given database.

notmuch_database_upgrade
Upgrade the current database to the latest supported version.

notmuch_directory_delete
Delete directory document from the database, and destroy the notmuch_directory_t object. Assumes any child directories and files have been deleted by the caller.

notmuch_directory_destroy
Destroy a notmuch_directory_t object.

notmuch_directory_get_child_directories
Get a notmuch_filenames_t iterator listing all the filenames of sub-directories in the database within the given directory.

notmuch_directory_get_child_files
Get a notmuch_filenames_t iterator listing all the filenames of messages in the database within the given directory.

notmuch_directory_get_mtime
Get the mtime of a directory, (as previously stored with notmuch_directory_set_mtime).

notmuch_directory_set_mtime
Store an mtime within the database for 'directory'.

notmuch_filenames_destroy
Destroy a notmuch_filenames_t object.

notmuch_filenames_get
Get the current filename from 'filenames' as a string.

notmuch_filenames_move_to_next
Move the 'filenames' iterator to the next filename.

notmuch_filenames_valid
Is the given 'filenames' iterator pointing at a valid filename.

notmuch_message_add_tag
Add a tag to the given message.

notmuch_message_destroy
Destroy a notmuch_message_t object.

notmuch_message_freeze
Freeze the current state of 'message' within the database.

notmuch_message_get_date
Get the date of 'message' as a time_t value.

notmuch_message_get_filename
Get a filename for the email corresponding to 'message'.

notmuch_message_get_filenames
Get all filenames for the email corresponding to 'message'.

notmuch_message_get_flag
Get a value of a flag for the email corresponding to 'message'.

notmuch_message_get_header
Get the value of the specified header from 'message' as a UTF-8 string.

notmuch_message_get_message_id
Get the message ID of 'message'.

notmuch_message_get_replies
Get a notmuch_messages_t iterator for all of the replies to 'message'.

notmuch_message_get_tags
Get the tags for 'message', returning a notmuch_tags_t object which can be used to iterate over all tags.

notmuch_message_get_thread_id
Get the thread ID of 'message'.

notmuch_message_maildir_flags_to_tags
Add/remove tags according to maildir flags in the message filename(s).

notmuch_message_remove_all_tags
Remove all tags from the given message.

notmuch_message_remove_tag
Remove a tag from the given message.

notmuch_message_set_flag
Set a value of a flag for the email corresponding to 'message'.

notmuch_message_tags_to_maildir_flags
Rename message filename(s) to encode tags as maildir flags.

notmuch_message_thaw
Thaw the current 'message', synchronizing any changes that may have occurred while 'message' was frozen into the notmuch database.

notmuch_messages_collect_tags
Return a list of tags from all messages.

notmuch_messages_destroy
Destroy a notmuch_messages_t object.

notmuch_messages_get
Get the current message from 'messages' as a notmuch_message_t.

notmuch_messages_move_to_next
Move the 'messages' iterator to the next message.

notmuch_messages_valid
Is the given 'messages' iterator pointing at a valid message.

notmuch_query_add_tag_exclude
Add a tag that will be excluded from the query results by default. This exclusion will be overridden if this tag appears explicitly in the query.

notmuch_query_count_messages_st
Return the number of messages matching a search.

notmuch_query_count_threads_st
Return the number of threads matching a search.

notmuch_query_create
Create a new query for 'database'.

notmuch_query_destroy
Destroy a notmuch_query_t along with any associated resources.

notmuch_query_get_database
Return the notmuch database of this query. See notmuch_query_create.

notmuch_query_get_query_string
Return the query_string of this query. See notmuch_query_create.

notmuch_query_get_sort
Return the sort specified for this query. See notmuch_query_set_sort.

notmuch_query_search_messages_st
Execute a query for messages, returning a notmuch_messages_t object which can be used to iterate over the results. The returned messages object is owned by the query and as such, will only be valid until notmuch_query_destroy.

notmuch_query_search_threads_st
Execute a query for threads, returning a notmuch_threads_t object which can be used to iterate over the results. The returned threads object is owned by the query and as such, will only be valid until notmuch_query_destroy.

notmuch_query_set_omit_excluded
Specify whether to omit excluded results or simply flag them. By default, this is set to TRUE.

notmuch_query_set_sort
Specify the sorting desired for this query.

notmuch_status_to_string
Get a string representation of a notmuch_status_t value.

notmuch_tags_destroy
Destroy a notmuch_tags_t object.

notmuch_tags_get
Get the current tag from 'tags' as a string.

notmuch_tags_move_to_next
Move the 'tags' iterator to the next tag.

notmuch_tags_valid
Is the given 'tags' iterator pointing at a valid tag.

notmuch_thread_destroy
Destroy a notmuch_thread_t object.

notmuch_thread_get_authors
Get the authors of 'thread' as a UTF-8 string.

notmuch_thread_get_matched_messages
Get the number of messages in 'thread' that matched the search.

notmuch_thread_get_messages
Get a notmuch_thread_t iterator for all messages in 'thread' in oldest-first order.

notmuch_thread_get_newest_date
Get the date of the newest message in 'thread' as a time_t value.

notmuch_thread_get_oldest_date
Get the date of the oldest message in 'thread' as a time_t value.

notmuch_thread_get_subject
Get the subject of 'thread' as a UTF-8 string.

notmuch_thread_get_tags
Get the tags for 'thread', returning a notmuch_tags_t object which can be used to iterate over all tags.

notmuch_thread_get_thread_id
Get the thread ID of 'thread'.

notmuch_thread_get_toplevel_messages
Get a notmuch_messages_t iterator for the top-level messages in 'thread' in oldest-first order.

notmuch_thread_get_total_messages
Get the total number of messages in 'thread'.

notmuch_threads_destroy
Destroy a notmuch_threads_t object.

notmuch_threads_get
Get the current thread from 'threads' as a notmuch_thread_t.

notmuch_threads_move_to_next
Move the 'threads' iterator to the next thread.

notmuch_threads_valid
Is the given 'threads' iterator pointing at a valid thread.

Type Definitions

notmuch_bool_t