Enum notmuch_sys::notmuch_message_flag_t

pub enum notmuch_message_flag_t {
    MATCH,
    EXCLUDED,
    GHOST,
}

Message flags.

Variants

MATCH

EXCLUDED

GHOST

This message is a "ghost message", meaning it has no filenames or content, but we know it exists because it was referenced by some other message. A ghost message has only a message ID and thread ID.