Enum notmuch_sys::notmuch_status_t

#[repr(C)]

pub enum notmuch_status_t {
    SUCCESS,
    OUT_OF_MEMORY,
    READ_ONLY_DATABASE,
    XAPIAN_EXCEPTION,
    FILE_ERROR,
    FILE_NOT_EMAIL,
    DUPLICATE_MESSAGE_ID,
    NULL_POINTER,
    TAG_TOO_LONG,
    UNBALANCED_FREEZE_THAW,
    UNBALANCED_ATOMIC,
    UNSUPPORTED_OPERATION,
    UPGRADE_REQUIRED,
    PATH_ERROR,
    NOTMUCH_STATUS_ILLEGAL_ARGUMENT,
}

Status codes used for the return values of most functions.

A zero value (SUCCESS) indicates that the function completed without error. Any other value indicates an error.

Variants
SUCCESS

No error occurred.

OUT_OF_MEMORY

Out of memory.

READ_ONLY_DATABASE

An attempt was made to write to a database opened in read-only mode.

XAPIAN_EXCEPTION

A Xapian exception occurred.

@todo We don't really want to expose this lame XAPIAN_EXCEPTION value. Instead we should map to things like DATABASE_LOCKED or whatever.

FILE_ERROR

An error occurred trying to read or write to a file (this could be file not found, permission denied, etc.)

FILE_NOT_EMAIL

A file was presented that doesn't appear to be an email message.

DUPLICATE_MESSAGE_ID

A file contains a message ID that is identical to a message already in the database.

NULL_POINTER

The user erroneously passed a NULL pointer to a notmuch function.

TAG_TOO_LONG

A tag value is too long (exceeds TAG_MAX).

UNBALANCED_FREEZE_THAW

The notmuch_message_thaw function has been called more times than notmuch_message_freeze.

UNBALANCED_ATOMIC

notmuch_database_end_atomic has been called more times than notmuch_database_begin_atomic.

UNSUPPORTED_OPERATION

The operation is not supported.

UPGRADE_REQUIRED

The operation requires a database upgrade.

PATH_ERROR

There is a problem with the proposed path, e.g. a relative path passed to a function expecting an absolute path.

NOTMUCH_STATUS_ILLEGAL_ARGUMENT

One of the arguments violates the preconditions for the function, in a way not covered by a more specific argument.