Function notmuch_sys::notmuch_database_compact

pub unsafe extern "C" fn notmuch_database_compact(
    path: *const c_char, 
    backup_path: *const c_char, 
    status_cb: Option<extern "C" fn(message: *const c_char, closure: *mut c_void)>, 
    closure: *mut c_void
) -> notmuch_status_t

Compact a notmuch database, backing up the original database to the given path.

The database will be opened with notmuch_database_mode_t::READ_WRITE during the compaction process to ensure no writes are made.

If the optional callback function 'status_cb' is non-NULL, it will be called with diagnostic and informational messages. The argument 'closure' is passed verbatim to any callback invoked.