notmuch_sys::notmuch_message_get_filename

pub unsafe extern fn notmuch_message_get_filename(message: *mut notmuch_message_t) -> *const c_char

Get a filename for the email corresponding to 'message'.

The returned filename is an absolute filename, (the initial component will match notmuch_database_get_path() ).

The returned string belongs to the message so should not be modified or freed by the caller (nor should it be referenced after the message is destroyed).

Note: If this message corresponds to multiple files in the mail store, (that is, multiple files contain identical message IDs), this function will arbitrarily return a single one of those filenames. See notmuch_message_get_filenames for returning the complete list of filenames.