notmuch_sys::notmuch_message_get_replies

pub unsafe extern fn notmuch_message_get_replies(message: *mut notmuch_message_t) -> *mut notmuch_messages_t

Get a notmuch_messages_t iterator for all of the replies to 'message'.

Note: This call only makes sense if 'message' was ultimately obtained from a notmuch_thread_t object, (such as by coming directly from the result of calling notmuch_thread_get_ toplevel_messages or by any number of subsequent calls to notmuch_message_get_replies).

If 'message' was obtained through some non-thread means, (such as by a call to notmuch_query_search_messages), then this function will return NULL.

If there are no replies to 'message', this function will return NULL. (Note that notmuch_messages_valid will accept that NULL value as legitimate, and simply return FALSE for it.)