notmuch_sys::notmuch_message_get_tags

pub unsafe extern fn notmuch_message_get_tags(message: *mut notmuch_message_t) -> *mut notmuch_tags_t

Get the tags for 'message', returning a notmuch_tags_t object which can be used to iterate over all tags.

The tags object is owned by the message and as such, will only be valid for as long as the message is valid, (which is until the query from which it derived is destroyed).

Typical usage might be:

notmuch_message_t *message;
notmuch_tags_t *tags;
const char *tag;

message = notmuch_database_find_message (database, message_id);

for (tags = `notmuch_message_get_tags` (message);
     notmuch_tags_valid (tags);
     notmuch_tags_move_to_next (tags))
{
    tag = notmuch_tags_get (tags);
    ....
}

notmuch_message_destroy (message);

Note that there's no explicit destructor needed for the notmuch_tags_t object. (For consistency, we do provide a notmuch_tags_destroy function, but there's no good reason to call it if the message is about to be destroyed).