notmuch_sys::notmuch_message_thaw

pub unsafe extern fn notmuch_message_thaw(message: *mut notmuch_message_t) -> notmuch_status_t

Thaw the current 'message', synchronizing any changes that may have occurred while 'message' was frozen into the notmuch database.

See notmuch_message_freeze for an example of how to use this function to safely provide tag changes.

Multiple calls to freeze/thaw are valid and these calls with "stack". That is there must be as many calls to thaw as to freeze before a message is actually thawed.

Return value:

notmuch_status_t::SUCCESS: Message successfully thawed, (or at least its frozen count has successfully been reduced by 1).

notmuch_status_t::UNBALANCED_FREEZE_THAW: An attempt was made to thaw an unfrozen message. That is, there have been an unbalanced number of calls to notmuch_message_freeze and notmuch_message_thaw.