notmuch_sys::notmuch_thread_get_matched_messages

pub unsafe extern fn notmuch_thread_get_matched_messages(thread: *mut notmuch_thread_t) -> c_int

Get the number of messages in 'thread' that matched the search.

This count includes only the messages in this thread that were matched by the search from which the thread was created and were not excluded by any exclude tags passed in with the query (see notmuch_query_add_tag_exclude). Contrast with notmuch_thread_get_total_messages() .